requirements.txt
Pillow==8.4.0
vkbottle==4.2.1
qrcode==7.3.1
pyowm==3.3.0
translate==3.6.1
